workflows: add workflow to recreate linux self-hosted VM on shedule
This workflow will run once each 24h to delete the linux runner and create a new one,
to increaste the security of the runner in case it is compromised.

This is a middle ground as we can't have disposable self-hosted runners.

Right now this PR is just a WIP that tries to understand how to fetch the list of runners
and their status from github.